week 1 solution
Solution for programming assignment 1: percolation. Consists of
two classes, Percolation and PercolationStats. Percolation accepts
a parameter n and generates an n x n size grid of sites. Sites can
be open or blocked, and a site that is open and connected to any
site on the top row is said to be full. The system is said to
percolate if there is a full site on the bottom row.
PercolationStats is a client of Percolation; it initializes a grid
and randomly opens sites until the system percolates. It performs
t trials in this manner and tracks the number of open sites at
which each simulation percolated, and reports statistics for the
simulations.