week 2 solution
Solution for programming assignment 2: queues. Consists of two
classes, Deque and RandomizedQueue, in addition to a client class
Permutation that uses RandomizedQueue. Deque is implemented as a
doubly-linked list; pointers to first and last nodes are kept and
updated as items are added or removed at the beginning or end.
Nodes must be doubly linked (i.e. track next and prev node) so
that all operations take constant time.
RandomizedQueue class is a resizing array implementation. When it
is full, it doubles in size; when it is 1/4 full, it halves. As a
queue it also has pointers to the first and next indices; when the
next pointer reaches the end of the array, it copies all items to
the beginning of the array and resets the pointers. The randomness
is encapsulated in the internal RandomizedQueueIterator; it
generates a random order and returns items in that order.
Permutation takes a command line arg k and then reads strings from
stdin, loads them into a RandomizedQueue, and prints k strings in
random order to stdout.