week 3 solution
Application consists of four classes: Point, LineSegment,
BruteCollinearPoints and FastCollinearPoints. LineSegment was
provided and Point was provided partially implemented. A Point
represents a point on the cartesian plane; a LineSegment links
two Points. BruteCollinearPoints is a brute-force algorithm that
reads input from stdin, generates points based on that input, and
finds all line segments on which four or more points lie.
FastCollinearPoints uses sorting to implement a faster solution:
we iterate over all points and for every point p, we sort the rest
of them by their sort order with respect to p. Since we know that
any points collinear with p will be adjacent to one another in the
sorted list, so we only have to search until we find such a group.
When we have found such a group, we know that we have a line that
contains at least four points, so we filter the points by the
slope of that line and sort them by their natural order in order
to find the maximal line segment (the line segment bounded by the
least and greatest points from the input that are found on that
line).